requirements.txt
requests
mastodon.py
pythorhead>=0.23.0
python-dotenv
loguru
Pillow
horde-sdk
loguru